Build out loader to load kernel successfully
This builds out the loader to load the kernel image from disk, build a set
of page tables for it, and starts to construct boot info to pass it. We
confirm success atm by printing to the debug console.

Further work will be needed to build a memory map from the UEFI memory map
to pass the kernel, from which it can bootstrap its own physical memory manager.

This commit introduces `mycelium-bitfield` as a dependency - we would very likely
build out very similar functionality ourselves and so this feels like a
reasonable external dependency to rely upon.